Simplify and split interface for executor services
This is done to simplify integration into canton.
